feat: use peak memory estimation from Han
